students/jonggyu/westagram/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
mysql-client==0.0.1
PyMySQL==1.0.2
django-phonenumber-field==5.2.0
phonenumbers==8.12.28
bcrypt==3.2.0
PyJWT==2.1.0
